refactor: split large components into smaller modules
- Split Popup.tsx (1063 lines) into main component + 7 sub-components
  - UIScaleSection, ImportExportSection, FirebaseAuthSection
  - SyncStatusSection, GameTypeFilterSection, HandLimitSection
  - StatisticsConfigSection
- Split Hud.tsx (681 lines) into main component + 6 sub-components
  - RealTimeStatsDisplay, HudHeader, StatDisplay
  - PlayerTypeIcons, DragHandle, useDraggable hook
- Organized sub-components in Popup/ and Hud/ directories
- Improved maintainability and testability
- No functional changes, all tests passing (181/181)
